Renamed whichkey folder to keybindings
