+++ init.lua
@@ -15,5 +15,5 @@ require("user.project")
 require("user.impatient")
 require("user.indentline")
 require("user.alpha")
-require("user.whichkey")
+require("user.keybindings")
 require("user.autocommands")
